update service worker to vite
