+++ public/sw.js
@@ -11,8 +11,27 @@ import { clientsClaim } from 'workbox-core';
 import { ExpirationPlugin } from 'workbox-expiration';
 import { precacheAndRoute, createHandlerBoundToURL } from 'workbox-precaching';
 import { registerRoute } from 'workbox-routing';
-import { StaleWhileRevalidate, NetworkFirst, CacheFirst } from 'workbox-strategies';
-import { CacheNetworkRace } from './sw-strategies';
+import { StaleWhileRevalidate, NetworkFirst, CacheFirst, Strategy } from 'workbox-strategies';
+
+class CacheNetworkRace extends Strategy {
+  _handle(request, handler) {
+      const fetchAndCachePutDone = handler.fetchAndCachePut(request)
+      const cacheMatchDone = handler.cacheMatch(request)
+
+      return new Promise((resolve, reject) => {
+          fetchAndCachePutDone.then(resolve)
+          cacheMatchDone.then(response => response && resolve(response))
+
+          // Reject if both network and cache error or find no response
+          Promise.allSettled([fetchAndCachePutDone, cacheMatchDone]).then(results => {
+              const [fetchAndCachePutResult, cacheMatchResult] = results
+              if (fetchAndCachePutResult.status === 'rejected' && !cacheMatchResult.value) {
+                  reject(fetchAndCachePutResult.reason)
+              }
+          })
+      })
+  }
+}
 
 clientsClaim();
 
@@ -21,31 +40,6 @@ clientsClaim();
 // This variable must be present somewhere in your service worker file,
 // even if you decide not to use precaching. See https://cra.link/PWA
 precacheAndRoute(self.__WB_MANIFEST);
-
-// Set up App Shell-style routing, so that all navigation requests
-// are fulfilled with your index.html shell. Learn more at
-// https://developers.google.com/web/fundamentals/architecture/app-shell
-const fileExtensionRegexp = new RegExp('/[^/?]+\\.[^/]+$');
-registerRoute(
-  // Return false to exempt requests from being fulfilled by index.html.
-  ({ request, url }) => {
-    // If this isn't a navigation, skip.
-    if (request.mode !== 'navigate') {
-      return false;
-    } // If this is a URL that starts with /_, skip.
-
-    if (url.pathname.startsWith('/_')) {
-      return false;
-    } // If this looks like a URL for a resource, because it contains // a file extension, skip.
-
-    if (url.pathname.match(fileExtensionRegexp)) {
-      return false;
-    } // Return true to signal that we want to use the handler.
-
-    return true;
-  },
-  createHandlerBoundToURL(import.meta.env.PUBLIC_URL + '/index.html')
-);
 
 // An example runtime caching route for requests that aren't handled by the
 // precache, in this case same-origin .png requests like those from in public/
